move message watch
