+++ pkg/pipeline/pipeline.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"os"
 	"path"
-	"regexp"
 	"sync"
 	"time"
 
@@ -29,15 +28,6 @@ const (
 	pipelineSource    = "pipeline"
 	eosTimeout        = time.Second * 30
 	maxPendingUploads = 100
-
-	fragmentOpenedMessage = "splitmuxsink-fragment-opened"
-	fragmentClosedMessage = "splitmuxsink-fragment-closed"
-	fragmentLocation      = "location"
-	fragmentRunningTime   = "running-time"
-
-	elementGstRtmp2Sink = "GstRtmp2Sink"
-	elementGstAppSrc    = "GstAppSrc"
-	elementSplitMuxSink = "GstSplitMuxSink"
 )
 
 type Pipeline struct {
@@ -279,104 +269,6 @@ func (p *Pipeline) Run(ctx context.Context) *livekit.EgressInfo {
 	return p.Info
 }
 
-func (p *Pipeline) messageWatch(msg *gst.Message) bool {
-	switch msg.Type() {
-	case gst.MessageEOS:
-		// EOS received - close and return
-		if p.eosTimer != nil {
-			p.eosTimer.Stop()
-		}
-
-		logger.Debugw("EOS received, stopping pipeline")
-		p.closeOnce.Do(func() {
-			p.close(context.Background())
-		})
-
-		p.stop()
-		return false
-
-	case gst.MessageError:
-		// handle error if possible, otherwise close and return
-		err, handled := p.handleError(msg.ParseError())
-		if !handled {
-			p.Info.Error = err.Error()
-			p.stop()
-			return false
-		}
-
-	case gst.MessageStateChanged:
-		if p.playing {
-			return true
-		}
-
-		_, newState := msg.ParseStateChanged()
-		if newState != gst.StatePlaying {
-			return true
-		}
-
-		switch msg.Source() {
-		case sdk.AudioAppSource, sdk.VideoAppSource:
-			switch s := p.in.(type) {
-			case *sdk.SDKInput:
-				s.Playing(msg.Source())
-			}
-
-		case pipelineSource:
-			p.playing = true
-			switch s := p.in.(type) {
-			case *sdk.SDKInput:
-				p.updateStartTime(s.GetStartTime())
-			case *web.WebInput:
-				p.updateStartTime(time.Now().UnixNano())
-			}
-		}
-
-	case gst.MessageElement:
-		s := msg.GetStructure()
-		if s != nil {
-			switch s.Name() {
-			case fragmentOpenedMessage:
-				filepath, t, err := getSegmentParamsFromGstStructure(s)
-				if err != nil {
-					logger.Errorw("failed to retrieve segment parameters from event", err)
-					return true
-				}
-
-				logger.Debugw("fragment opened", "location", filepath, "running time", t)
-
-				if p.playlistWriter != nil {
-					if err = p.playlistWriter.StartSegment(filepath, t); err != nil {
-						logger.Errorw("failed to register new segment with playlist writer", err, "location", filepath, "running time", t)
-						return true
-					}
-				}
-
-			case fragmentClosedMessage:
-				filepath, t, err := getSegmentParamsFromGstStructure(s)
-				if err != nil {
-					logger.Errorw("failed to retrieve segment parameters from event", err, "location", filepath, "running time", t)
-					return true
-				}
-
-				logger.Debugw("fragment closed", "location", filepath, "running time", t)
-
-				// We need to dispatch to a queue to:
-				// 1. Avoid concurrent access to the SegmentsInfo structure
-				// 2. Ensure that playlists are uploaded in the same order they are enqueued to avoid an older playlist overwriting a newer one
-				if err = p.enqueueSegmentUpload(filepath, t); err != nil {
-					logger.Errorw("failed to end segment with playlist writer", err, "running time", t)
-					return true
-				}
-			}
-		}
-
-	default:
-		logger.Debugw(msg.String())
-	}
-
-	return true
-}
-
 func (p *Pipeline) UpdateStream(ctx context.Context, req *livekit.UpdateStreamRequest) error {
 	ctx, span := tracer.Start(ctx, "Pipeline.UpdateStream")
 	defer span.End()
@@ -420,11 +312,6 @@ func (p *Pipeline) UpdateStream(ctx context.Context, req *livekit.UpdateStreamRe
 	return errs.ToError()
 }
 
-func (p *Pipeline) GetGstPipelineDebugDot() (string, error) {
-	s := p.pipeline.DebugBinToDotData(gst.DebugGraphShowAll)
-	return s, nil
-}
-
 func (p *Pipeline) removeSink(url string, status livekit.StreamInfo_Status) error {
 	now := time.Now().UnixNano()
 
@@ -460,6 +347,11 @@ func (p *Pipeline) removeSink(url string, status livekit.StreamInfo_Status) erro
 	return p.out.RemoveSink(url)
 }
 
+func (p *Pipeline) GetGstPipelineDebugDot() (string, error) {
+	s := p.pipeline.DebugBinToDotData(gst.DebugGraphShowAll)
+	return s, nil
+}
+
 func (p *Pipeline) SendEOS(ctx context.Context) {
 	ctx, span := tracer.Start(ctx, "Pipeline.SendEOS")
 	defer span.End()
@@ -485,23 +377,6 @@ func (p *Pipeline) SendEOS(ctx context.Context) {
 	})
 }
 
-func (p *Pipeline) close(ctx context.Context) {
-	close(p.closed)
-	if p.limitTimer != nil {
-		p.limitTimer.Stop()
-	}
-
-	switch p.Info.Status {
-	case livekit.EgressStatus_EGRESS_STARTING,
-		livekit.EgressStatus_EGRESS_ACTIVE:
-
-		p.Info.Status = livekit.EgressStatus_EGRESS_ENDING
-		if p.onStatusUpdate != nil {
-			p.onStatusUpdate(ctx, p.Info)
-		}
-	}
-}
-
 func (p *Pipeline) startSessionLimitTimer(ctx context.Context) {
 	var timeout time.Duration
 
@@ -519,31 +394,6 @@ func (p *Pipeline) startSessionLimitTimer(ctx context.Context) {
 			p.SendEOS(ctx)
 			p.Info.Status = livekit.EgressStatus_EGRESS_LIMIT_REACHED
 		})
-	}
-}
-
-func (p *Pipeline) updateStartTime(startedAt int64) {
-	switch p.EgressType {
-	case types.EgressTypeStream, types.EgressTypeWebsocket:
-		p.mu.Lock()
-		for _, streamInfo := range p.StreamInfo {
-			streamInfo.Status = livekit.StreamInfo_ACTIVE
-			streamInfo.StartedAt = startedAt
-		}
-		p.mu.Unlock()
-
-	case types.EgressTypeFile:
-		p.FileInfo.StartedAt = startedAt
-
-	case types.EgressTypeSegmentedFile:
-		p.SegmentsInfo.StartedAt = startedAt
-	}
-
-	if p.Info.Status == livekit.EgressStatus_EGRESS_STARTING {
-		p.Info.Status = livekit.EgressStatus_EGRESS_ACTIVE
-		if p.onStatusUpdate != nil {
-			p.onStatusUpdate(context.Background(), p.Info)
-		}
 	}
 }
 
@@ -599,6 +449,31 @@ func (p *Pipeline) enqueueSegmentUpload(segmentPath string, endTime int64) error
 	}
 }
 
+func (p *Pipeline) updateStartTime(startedAt int64) {
+	switch p.EgressType {
+	case types.EgressTypeStream, types.EgressTypeWebsocket:
+		p.mu.Lock()
+		for _, streamInfo := range p.StreamInfo {
+			streamInfo.Status = livekit.StreamInfo_ACTIVE
+			streamInfo.StartedAt = startedAt
+		}
+		p.mu.Unlock()
+
+	case types.EgressTypeFile:
+		p.FileInfo.StartedAt = startedAt
+
+	case types.EgressTypeSegmentedFile:
+		p.SegmentsInfo.StartedAt = startedAt
+	}
+
+	if p.Info.Status == livekit.EgressStatus_EGRESS_STARTING {
+		p.Info.Status = livekit.EgressStatus_EGRESS_ACTIVE
+		if p.onStatusUpdate != nil {
+			p.onStatusUpdate(context.Background(), p.Info)
+		}
+	}
+}
+
 func (p *Pipeline) updateDuration(endedAt int64) {
 	p.mu.Lock()
 	defer p.mu.Unlock()
@@ -627,6 +502,23 @@ func (p *Pipeline) updateDuration(endedAt int64) {
 		}
 		p.SegmentsInfo.EndedAt = endedAt
 		p.SegmentsInfo.Duration = endedAt - p.SegmentsInfo.StartedAt
+	}
+}
+
+func (p *Pipeline) close(ctx context.Context) {
+	close(p.closed)
+	if p.limitTimer != nil {
+		p.limitTimer.Stop()
+	}
+
+	switch p.Info.Status {
+	case livekit.EgressStatus_EGRESS_STARTING,
+		livekit.EgressStatus_EGRESS_ACTIVE:
+
+		p.Info.Status = livekit.EgressStatus_EGRESS_ENDING
+		if p.onStatusUpdate != nil {
+			p.onStatusUpdate(ctx, p.Info)
+		}
 	}
 }
 
@@ -746,85 +638,4 @@ func (p *Pipeline) cleanup() {
 			}
 		}
 	}
-}
-
-func getSegmentParamsFromGstStructure(s *gst.Structure) (filepath string, time int64, err error) {
-	loc, err := s.GetValue(fragmentLocation)
-	if err != nil {
-		return "", 0, err
-	}
-	filepath, ok := loc.(string)
-	if !ok {
-		return "", 0, errors.ErrGstPipelineError(errors.New("invalid type for location"))
-	}
-
-	t, err := s.GetValue(fragmentRunningTime)
-	if err != nil {
-		return "", 0, err
-	}
-	ti, ok := t.(uint64)
-	if !ok {
-		return "", 0, errors.ErrGstPipelineError(errors.New("invalid type for time"))
-	}
-
-	return filepath, int64(ti), nil
-}
-
-// handleError returns true if the error has been handled, false if the pipeline should quit
-func (p *Pipeline) handleError(gErr *gst.GError) (error, bool) {
-	element, name, message := parseDebugInfo(gErr)
-	err := errors.ErrGstPipelineError(errors.New(gErr.Error()))
-
-	switch {
-	case element == elementGstRtmp2Sink:
-		// bad URI or could not connect. Remove rtmp output
-		url, e := p.out.GetUrlFromName(name)
-		if e != nil {
-			logger.Warnw("rtmp output not found", e, "url", url)
-			return e, false
-		}
-		if e = p.removeSink(url, livekit.StreamInfo_FAILED); e != nil {
-			return err, false
-		}
-		return err, true
-
-	case element == elementGstAppSrc:
-		if message == "streaming stopped, reason not-negotiated (-4)" {
-			// send eos to app src
-			logger.Debugw("streaming stopped", "name", name)
-			p.in.(*sdk.SDKInput).StreamStopped(name)
-			return err, true
-		}
-	case element == elementSplitMuxSink:
-		// We sometimes get GstSplitMuxSink errors if send EOS before the first media was sent to the mux
-		if message == ":muxer" {
-			select {
-			case <-p.closed:
-				logger.Debugw("GstSplitMuxSink failure after sending EOS")
-				return err, true
-			default:
-			}
-		}
-	}
-
-	// input failure or file write failure. Fatal
-	logger.Errorw("pipeline error", err,
-		"element", element,
-		"message", message,
-	)
-
-	return err, false
-}
-
-// Debug info comes in the following format:
-// file.c(line): method_name (): /GstPipeline:pipeline/GstBin:bin_name/GstElement:element_name:\nError message
-var regExp = regexp.MustCompile("(?s)(.*?)GstPipeline:pipeline\\/GstBin:(.*?)\\/(.*?):([^:]*)(:\n)?(.*)")
-
-func parseDebugInfo(gErr *gst.GError) (element, name, message string) {
-	match := regExp.FindStringSubmatch(gErr.DebugString())
-
-	element = match[3]
-	name = match[4]
-	message = match[6]
-	return
 }
